backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
celery>=5.3.1
redis>=4.6.0
pydantic-settings>=2.0.3
websockets>=11.0.3
httpx>=0.24.1
aiofiles>=23.2.1
ffmpeg-python>=0.2.0
edge-tts>=6.1.12
